This is the first edition
This is a Kenyan meme drum kit. If you click on a button or press a certain key, it plays an audio from popular Kenyan video. It would be quite hilarious as a quick witted quip or reply during a zoom meeting. As you can probably guess, this only makes sense in the context of Kenyan social media circles in 2022/2023.